Seventh build. Removed useless code, reworked the image validation by type, corrected the syntax to the programming standards, secured the delete function and prevented non admin users from deleting, prevented sql injection with parametrization and string escaping and remade the "adding product" function to not accept invalid data.

+++ web/delete.php
@@ -1,14 +1,21 @@
 <?php
 require_once("common.php");
+if(isset($_SESSION['admin'])){
+    if(isset($_GET['id'])){
+        $stmt = "";
+        $id = sanitize($_GET['id']);
+        if(!($stmt = $link->prepare("delete from products where id=(?)"))){
+            echo "Update Prepare error ".$link->error;
+        }
+        if(!($stmt->bind_param('d',$id))){
+            echo "Update Bind error ".$link->error;
+        }
+        $result = getProduct($id);
+        unlink($result['image']);
 
-if(isset($_GET['id'])){
-    $option = "delete from products where id=".$_GET['id'].";";
-
-    $result = getProduct($_GET['id']);
-    unlink($result['image']);
-
-    if(!mysqli_query($link,$option)){
-        echo mysqli_error($link);
+        if (!($stmt->execute())) {
+            echo "Execute error " . $link->error;
+        }
     }
 }
 header("Location:admin.php");
